+++ CHANGELOG.md
@@ -40,6 +40,8 @@ version mostly conforms to the state as documented in the ISLa paper.
 - Factored out functions related to ISLa evaluations from `isla.language` to `isla.evaluator`.
   The previous `isla.evaluator` file was named `isla.performance_evaluator`.
 - Corrections in README.txt
+- Checking whether existential (integer / tree) quantifiers are already satisfied, taking into account
+  existing constraints (assumptions).
 
 ### Removed